backend/requirements.txt
Django>=4.0,<5.0

# REST API & auth
djangorestframework>=3.12
django-rest-authtoken
django-cors-headers>=3.7
django-allauth>=0.51
dj-rest-auth>=2.2

# Database driver for PostgreSQL
psycopg2-binary>=2.8,<3.0

# HTTP requests & env
requests>=2.26
python-dotenv>=0.19

# testing
pytest>=7.0
pytest-django>=4.5
